# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ohio Casino API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rounds:
    post:
      operationId: placeBet
      tags: [rounds]
      summary: Place a bet and start a game round
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BetRequest"
            example:
              gameType: "slots"
              betAmount: 1000
              clientSeed: "a3f7b2d98f3c2a1b..."
              gameParams:
                demo: false
      responses:
        "200":
          description: Bet placed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetResponse"
              example:
                roundId: 12345
                serverSeedHash: "a1b2c3d4e5f6..."
                result: "pending"
                payout: 0
                newBalance: 9500
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Bet amount exceeds maximum"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Unauthorized"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Internal server error"
    get:
      operationId: listRounds
      tags: [rounds]
      summary: List game rounds for the authenticated user
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of rounds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameRound"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /rounds/{id}:
    get:
      operationId: getRound
      tags: [rounds]
      summary: Get game round details
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Round details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameRound"
              example:
                id: 12345
                userId: 987
                gameType: "slots"
                betAmount: 1000
                payout: 5000
                result: "win"
                serverSeedHash: "a1b2c3d4e5f6..."
                clientSeed: "my-client-seed-123"
                nonce: 42
                verified: true
                createdAt: "2025-06-23T10:00:00Z"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Unauthorized"
        "404":
          description: Round not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Round not found"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Internal server error"
  /rounds/{id}/verify:
    post:
      operationId: verifyRound
      tags: [rounds]
      summary: Verify a game round's fairness
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationRequest"
            example:
              roundId: 12345
              serverSeed: "slots:a3f7b2d9:42:8f3c2a1b..."
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResponse"
              example:
                verified: true
                computedHash: "a1b2c3d4e5f6..."
                expectedHash: "a1b2c3d4e5f6..."
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Invalid request"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Unauthorized"
        "404":
          description: Round not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Round not found"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Internal server error"
  /wallet:
    get:
      operationId: getWallet
      tags: [wallet]
      summary: Get the authenticated user's wallet
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Wallet details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
              example:
                id: 1
                userId: 987
                balance: 9500
                currency: "USD"
                createdAt: "2025-01-15T08:00:00Z"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Unauthorized"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Internal server error"
  /wallet/history:
    get:
      operationId: getWalletHistory
      tags: [wallet]
      summary: Get wallet transaction history
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Transaction history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
              example:
                - id: 1
                  walletId: 1
                  userId: 987
                  type: "bet"
                  amount: -1000
                  balanceBefore: 10500
                  balanceAfter: 9500
                  status: "completed"
                  referenceId: "round_12345"
                  description: "Bet on slots"
                  createdAt: "2025-06-23T10:00:00Z"
                - id: 2
                  walletId: 1
                  userId: 987
                  type: "win"
                  amount: 5000
                  balanceBefore: 9500
                  balanceAfter: 14500
                  status: "completed"
                  referenceId: "round_12345"
                  description: "Win on slots"
                  createdAt: "2025-06-23T10:01:00Z"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Unauthorized"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Internal server error"
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    GameRound:
      type: object
      required:
        [
          id,
          userId,
          gameType,
          betAmount,
          payout,
          serverSeedHash,
          clientSeed,
          nonce,
          verified,
          createdAt,
        ]
      properties:
        id:
          type: integer
        userId:
          type: integer
        gameType:
          type: string
        betAmount:
          type: integer
          description: Bet amount in cents
        payout:
          type: integer
          description: Payout in cents
        result:
          type: ["string", "null"]
        serverSeedHash:
          type: string
        clientSeed:
          type: string
        nonce:
          type: integer
        verified:
          type: boolean
        createdAt:
          type: string
    Wallet:
      type: object
      required: [id, userId, balance, currency, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        balance:
          type: integer
          description: Balance in cents
        currency:
          type: string
        createdAt:
          type: string
    BetRequest:
      type: object
      required: [gameType, betAmount, clientSeed]
      properties:
        gameType:
          type: string
          example: "slots"
        betAmount:
          type: integer
          minimum: 1
          description: Bet amount in cents
          example: 1000
        clientSeed:
          type: string
          example: "a3f7b2d98f3c2a1b..."
        gameParams:
          type: object
          description: Optional game-specific parameters
          example:
            demo: false
            rows: 16
            risk: "medium"
    BetResponse:
      type: object
      required: [roundId, serverSeedHash, result, payout, newBalance]
      properties:
        roundId:
          type: integer
        serverSeedHash:
          type: string
        result:
          type: string
        payout:
          type: integer
          description: Payout in cents
        newBalance:
          type: integer
          description: New wallet balance in cents
...
    VerificationRequest:
      type: object
      required: [roundId, serverSeed]
      properties:
        roundId:
          type: integer
          example: 12345
        serverSeed:
          type: string
          example: "slots:a3f7b2d9:42:8f3c2a1b..."
    VerificationResponse:
      type: object
      required: [verified, computedHash, expectedHash]
      properties:
        verified:
          type: boolean
          example: true
        computedHash:
          type: string
          example: "a1b2c3d4e5f6..."
        expectedHash:
          type: string
          example: "a1b2c3d4e5f6..."
...
    Transaction:
      type: object
      required:
        [
          id,
          walletId,
          userId,
          type,
          amount,
          balanceBefore,
          balanceAfter,
          status,
          createdAt,
        ]
      properties:
        id:
          type: integer
          example: 1
        walletId:
          type: integer
          example: 1
        userId:
          type: integer
          example: 987
        type:
          type: string
          example: "bet"
        amount:
          type: integer
          description: Amount in cents
          example: -1000
        balanceBefore:
          type: integer
          description: Balance before in cents
          example: 10500
        balanceAfter:
          type: integer
          description: Balance after in cents
          example: 9500
        status:
          type: string
          example: "completed"
        referenceId:
          type: ["string", "null"]
          example: "round_12345"
        description:
          type: ["string", "null"]
          example: "Bet on slots"
        createdAt:
          type: string
          example: "2025-06-23T10:00:00Z"